requirements.txt
python-dotenv==1.0.1
langchain==0.1.13
langchain-community==0.0.31
pinecone-client==3.2.2
openai==1.14.3
tiktoken==0.6.0
qdrant_client==1.8.2